adjust input spacing, add icons #15
