+++ client/src/styles/App.scss
@@ -1,7 +1,3 @@
 @import 'mixins';
 @import 'variables';
 @import 'register';
-
-body {
-    background-color: green;
-}
